Mismatching typedef int ObjectID. All the other places use size_t, so I changed this one to size_t too.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6155 3db68df2-c116-0410-a063-a993310a9797

+++ source/tools/atlas/AtlasUI/ScenarioEditor/Tools/Common/MiscState.h
@@ -5,7 +5,7 @@
 
 namespace AtlasMessage
 {
-	typedef int ObjectID;
+	typedef size_t ObjectID;
 }
 
 extern wxString g_SelectedTexture;
